Move VideoProfile into ffmpeg package.
Avoids a circular dependency between ffmpeg and core once ffmpeg
starts using the profiles.
