+++ ffmpeg/videoprofile.go
@@ -1,4 +1,4 @@
-package core
+package ffmpeg
 
 import (
 	"strconv"
